+++ test/dp_exchange/coinbase/baseline_test.exs
@@ -33,7 +33,9 @@ defmodule DpExchange.Coinbase.BaselineTest do
       credentials: nil,
       delivering: MapSet.new(),
       connected_once?: false,
-      last_seq: nil
+      last_seq: nil,
+      last_frame_at: nil,
+      silence_check: nil
     }
 
   defp frame(payload), do: Socket.handle_frame({:text, Jason.encode!(payload)}, state())